Refactor some things and upversion everything
