+++ cardDatabase/templates/cardDatabase/html/database_base.html
@@ -39,8 +39,8 @@
             <div id="basic-search">
                 <form id="basic-form" method="post" action="{% url 'cardDatabase-search' %}">
                     {{ basic_form }}
-                    <div class="form-submit-button">
-                        <input type="submit" value="Search" name="basic-form">
+                    <div >
+                        <input class="form-submit-button" type="submit" value="Search" name="basic-form">
                     </div>
                     {% include 'cardDatabase/html/search_tooltip.html' %}
                 </form>
