+++ cardDatabase/views/mobile_only.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def get(request):
+    return render(request, 'cardDatabase/html/mobile_only.html', {})
