+++ cardDatabase/views/pack_history.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def get(request):
+ return render(request, 'cardDatabase/html/pack_history.html', {})
