+++ cardDatabase/views/private_decklist.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def get(request):
+    return render(request, 'cardDatabase/html/private_decklist.html')
