+++ cardDatabase/views/user_preferences.py
@@ -1,0 +1,5 @@
+from django.http import HttpResponse
+
+
+def get(request):
+    return HttpResponse('')
